Major victory: ge.ml now has NO staging annotations AT ALL.  It is only code that concerns GE.
The code in infra.ml contains all the annotations.  A lot of it could be done automatically [but not in MetaOCaml itself, as modules cannot be manipulated as first-class values].  But a little script in your favourite untyped language would work ;-)

Had to put in an 'instantiate' function in the tests too.

An interesting side-effect is that somehow this different instantiation gets a different environment each time, which means that the counts for variable name generation gets reset each time, which means that variables names are now much shorter.