re-enable commented out tests fix code so that they work too!
This uncovered a bug where in the FractionFree case (but where the determinant was not output), there was a mismatch in the inputs.  Some of the changes in the types actually *caught* that!  I don't know how this managed to slip through before.  Unfortunately, none of the actual tests changed because of that.

Fix up Pivot type to be like the rest (almost)