Functorize the hell out of everything!
There is a chance I have gone overboard in the parametrization here...  I am not so sure that Infra should be a *functor*, perhaps having it be just a module would be OK.  And then (to make the types work out), having GEMake take an Infra module would be quite sufficient.  It would probably make things compile faster!