Almost completely integrated infrastructure for the full LU family to co-exist with GE.  This required more work on pivoting, then work on 'tracking' L (instead of just U, and more to the point, not obliterating L).
Use optSeqM in appropriate places -- this ensures the code does not contain useless ().

One can now output LU, but the code is not fully finished as it really outputs a combined L and U (twice).  Really the code needs to 'declare' a new L matrix and track the information there, or put the information in U and take it out into L before output.  Also need to implement a "packed" output (a la NAG).